refactor(client): http utils paths
