+++ packages/client/src/utils/HTTP/index.tsx
@@ -1,4 +1,4 @@
 export { buildPath } from './buildPath';
-export { determineAPIHost } from './determineAPIHost';
+export { determineAPIHost } from './determineHost';
 export { type ResponseType, HTTP } from './HTTP';
 export { HTTPClient } from './httpClient';
